+++ tests/test_responses.py
@@ -1,7 +1,9 @@
 import datetime as dt
 import os
+import sys
 import time
 import typing
+from asyncio import Queue
 from http.cookies import SimpleCookie
 
 import anyio
@@ -306,6 +308,43 @@ def test_file_response(tmpdir, test_client_factory):
     assert "last-modified" in response.headers
     assert "etag" in response.headers
     assert filled_by_bg_task == "6, 7, 8, 9"
+
+
+@pytest.mark.parametrize(
+    "extensions,result",
+    [
+        ({"http.response.pathsend": {}}, "http.response.pathsend"),
+        ({"http.response.zerocopysend": {}}, "http.response.zerocopysend"),
+    ],
+)
+async def test_file_response_optimizations(tmpdir, extensions, result, anyio_backend):
+    if sys.version_info < (3, 10) and anyio_backend == "trio":
+        pytest.skip("Not supported combination of trio, python  < 3.10 and asyncio.Queue")
+    path = os.path.join(tmpdir, "xyz")
+    content = b"<file content>" * 1000
+    with open(path, "wb") as file:
+        file.write(content)
+
+    fresponse = FileResponse(path=path, filename="example.png")
+    fresponse.chunk_size = 10
+    responses = Queue()
+    await fresponse({"extensions": extensions, "type": "response"}, None, responses.put)
+    response1 = await responses.get()
+    response2 = await responses.get()
+
+    expected_disposition = 'attachment; filename="example.png"'
+    assert response1["headers"]["content-type"] == "image/png"
+    assert response1["headers"]["content-disposition"] == expected_disposition
+    assert response2["type"] == result
+    if result == "http.response.pathsend":
+        assert response2["path"] == path
+    else:
+        assert response2["count"] == fresponse.chunk_size
+        assert response2["file"] > 0
+        assert response2["more_body"]
+        while response2["more_body"]:
+            response2 = await responses.get()
+        assert not response2["more_body"]
 
 
 def test_file_response_with_directory_raises_error(tmpdir, test_client_factory):